planets_api/planets_connector/serializers.py
from rest_framework import serializers
from planets_connector.models import Planet


class PlanetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Planet
        fields = '__all__'
