requirements.txt
Django==5.2.3
djangorestframework==3.16.0
requests==2.32.4
django-extensions==4.1